requirements.txt
requests==2.32.3
streamlit==1.44.1
pytest==8.3.5
python-dotenv==1.0.1
